Extract talpa form batch footer as it'sown component
