+++ frontend/benefit/shared/src/constants.ts
@@ -142,6 +142,8 @@ export enum APPLICATION_STATUSES {
 export enum BATCH_STATUSES {
   DRAFT = 'draft',
   AWAITING_FOR_DECISION = 'awaiting_ahjo_decision',
+  DECIDED_ACCEPTED = 'accepted_ahjo_decision',
+  DECIDED_REJECTED = 'rejected_ahjo_decision',
   SENT_TO_TALPA = 'sent_to_talpa',
 }
